changes to get rid of warning in Win V2017

+++ src/backend/base/camera_device.cpp
@@ -67,7 +67,7 @@ namespace bias
         return false;
     }
 
-    unsigned int CameraDevice::getNumberOfImageMode()
+    size_t CameraDevice::getNumberOfImageMode()
     {
         return 0;
     }
