+++ src/backend/spin/camera_device_spin.hpp
@@ -53,7 +53,7 @@ namespace bias {
             
             virtual bool isSupported(VideoMode vidMode, FrameRate frmRate);
             virtual bool isSupported(ImageMode imgMode);
-            virtual unsigned int getNumberOfImageMode();
+            virtual size_t getNumberOfImageMode();
 
             virtual VideoMode getVideoMode();
             virtual FrameRate getFrameRate();
