+++ src/utility/mat_to_qimage.cpp
@@ -14,9 +14,9 @@ namespace bias
             const uchar *qImageBuffer = (const uchar*) mat.data;
             QImage img = QImage(
                     qImageBuffer, 
-                    mat.cols, 
-                    mat.rows, 
-                    mat.step, 
+                    static_cast<int>(mat.cols), 
+                    static_cast<int>(mat.rows), 
+                    static_cast<int>(mat.step), 
                     QImage::Format_Indexed8
                     );
 
@@ -31,9 +31,9 @@ namespace bias
             const uchar *qImageBuffer = (const uchar*)mat.data;
             QImage img = QImage(
                     qImageBuffer, 
-                    matBGR.cols, 
-                    matBGR.rows, 
-                    matBGR.step, 
+                    static_cast<int>(matBGR.cols), 
+                    static_cast<int>(matBGR.rows), 
+                    static_cast<int>(matBGR.step), 
                     QImage::Format_RGB888
                     );
             return img.rgbSwapped();
@@ -43,9 +43,9 @@ namespace bias
             const uchar *qImageBuffer = (const uchar*)mat.data;
             QImage img = QImage(
                     qImageBuffer, 
-                    mat.cols, 
-                    mat.rows, 
-                    mat.step, 
+                    static_cast<float>(mat.cols), 
+                    static_cast<float>(mat.rows), 
+                    static_cast<float>(mat.step), 
                     QImage::Format_RGB888
                     );
             return img.rgbSwapped();
